Move non-golden test to cardano-cli-test
