+++ cardano-cli/cardano-cli.cabal
@@ -297,6 +297,7 @@ test-suite cardano-cli-test
                       , hedgehog
                       , hedgehog-extras ^>= 0.6.1.0
                       , exceptions
+                      , regex-tdfa
                       , tasty
                       , tasty-hedgehog
                       , text
