+++ cardano-cli/test/cardano-cli-test/Test/Cli/Governance/Hash.hs
@@ -1,8 +1,10 @@
 {- HLINT ignore "Use camelCase" -}
+{-# LANGUAGE TypeApplications #-}
 
 module Test.Cli.Governance.Hash where
 
-import           Control.Monad (void)
+import           Control.Monad (forM_, void)
+import           Text.Regex.TDFA ((=~))
 
 import           Test.Cardano.CLI.Util
 
@@ -40,3 +42,26 @@ governance_hash_trip_fun input =
     hashFromFile <- H.readFile hashFile
 
     H.diff hash (==) hashFromFile
+
+-- | Execute me with:
+-- @cabal test cardano-cli-test --test-options '-p "/governance committee key hash/"'@
+hprop_governance_committee_key_hash :: Property
+hprop_governance_committee_key_hash =
+  let supplyValues = [ "key-gen-cold", "key-gen-hot" ] in
+  propertyOnce $ forM_ supplyValues $ \flag ->
+    H.moduleWorkspace "tmp" $ \tempDir -> do
+    verificationKeyFile <- noteTempFile tempDir "key-gen.vkey"
+    signingKeyFile <- noteTempFile tempDir "key-gen.skey"
+
+    void $ execCardanoCLI
+      [ "conway", "governance", "committee", flag
+      , "--verification-key-file", verificationKeyFile
+      , "--signing-key-file", signingKeyFile
+      ]
+
+    result <- execCardanoCLI
+      [  "conway", "governance", "committee", "key-hash"
+      , "--verification-key-file", verificationKeyFile
+      ]
+
+    H.assert $ result =~ id @String "^[a-f0-9]{56}$"
